no results screen and clear search button
