+++ package.json
@@ -16,6 +16,7 @@
     "@types/react-dom": "^16.9.5",
     "@types/react-highlight-words": "^0.16.1",
     "@types/react-infinite-scroller": "^1.2.1",
+    "@types/react-linkify": "^1.0.0",
     "emotion-theming": "^10.0.27",
     "moment-timezone": "^0.5.28",
     "react": "^16.13.0",
