+++ src/App.tsx
@@ -2,11 +2,11 @@ import React, { useState } from 'react';
 import { Box, CSSReset, Grid, ThemeProvider } from '@chakra-ui/core';
 import InfiniteScroll from 'react-infinite-scroller';
 
-import { Filter, Loading, Meeting } from './components';
+import { Filter, Loading, Meeting, NoResults } from './components';
 import {
-  endpointUrl,
-  filterData,
-  loadStateFromResult,
+  endpoint,
+  filter,
+  load,
   meetingsPerPage,
   State,
   setQuery
@@ -27,12 +27,27 @@ export default function App() {
     timezone: ''
   });
 
+  //function to remove a tag
+  const toggleTag = (filter: string, value: string, checked: boolean): void => {
+    //loop through and add the tag
+    state.filters[filter].forEach(tag => {
+      if (tag.tag === value) {
+        tag.checked = checked;
+      } else if (['days', 'formats'].includes(filter)) {
+        //if we're setting a tag or format, uncheck the others
+        tag.checked = false;
+      }
+    });
+    //this will cause a re-render; the actual filtering is done in filterData
+    setState({ ...state });
+  };
+
   if (state.loading) {
     //on first render, get data
-    fetch(endpointUrl('1tYV4wBZkY_3hp0tresN6iZBCwOyqkK-dz4UAWQPI1Vs'))
+    fetch(endpoint('1tYV4wBZkY_3hp0tresN6iZBCwOyqkK-dz4UAWQPI1Vs'))
       .then(result => result.json())
       .then(result => {
-        setState(loadStateFromResult(result));
+        setState(load(result));
       });
   } else {
     //on subsequent renders, set query string
@@ -48,7 +63,7 @@ export default function App() {
     })
     .flat();
 
-  const filteredMeetings = filterData(state, tags);
+  const filteredMeetings = filter(state, tags);
 
   return (
     <ThemeProvider>
@@ -64,53 +79,40 @@ export default function App() {
           >
             <Box as="section" order={{ xs: 1, md: 2 }}>
               <Filter
-                filters={state.filters}
                 setSearch={(search: string[]) => {
                   setState({ ...state, search });
                 }}
                 setTimezone={(timezone: string) => {
                   setState({ ...state, timezone });
                 }}
-                toggleTag={(
-                  filter: string,
-                  value: string,
-                  checked: boolean
-                ) => {
-                  //loop through and add the tag
-                  state.filters[filter].forEach(tag => {
-                    if (tag.tag === value) {
-                      tag.checked = checked;
-                    } else if (['days', 'formats'].includes(filter)) {
-                      //if we're setting a tag or format, uncheck the others
-                      tag.checked = false;
-                    }
-                  });
-
-                  //this will cause a re-render; the actual filtering is done in filterData
-                  setState({ ...state });
-                }}
-                timezone={state.timezone}
+                state={state}
+                toggleTag={toggleTag}
               />
             </Box>
             <Box order={{ xs: 2, md: 1 }} overflow="hidden">
-              <InfiniteScroll
-                loadMore={() => {
-                  const limit = state.limit + meetingsPerPage;
-                  setState({ ...state, limit });
-                }}
-                hasMore={filteredMeetings.length > state.limit}
-              >
-                {filteredMeetings
-                  .slice(0, state.limit)
-                  .map((meeting: Meeting, index: number) => (
-                    <Meeting
-                      key={index}
-                      meeting={meeting}
-                      search={state.search}
-                      tags={tags}
-                    />
-                  ))}
-              </InfiniteScroll>
+              {!filteredMeetings.length && (
+                <NoResults state={state} toggleTag={toggleTag} />
+              )}
+              {!!filteredMeetings.length && (
+                <InfiniteScroll
+                  loadMore={() => {
+                    const limit = state.limit + meetingsPerPage;
+                    setState({ ...state, limit });
+                  }}
+                  hasMore={filteredMeetings.length > state.limit}
+                >
+                  {filteredMeetings
+                    .slice(0, state.limit)
+                    .map((meeting: Meeting, index: number) => (
+                      <Meeting
+                        key={index}
+                        meeting={meeting}
+                        search={state.search}
+                        tags={tags}
+                      />
+                    ))}
+                </InfiniteScroll>
+              )}
             </Box>
           </Grid>
         </Box>
