+++ src/components/Filter.tsx
@@ -3,36 +3,29 @@ import moment from 'moment-timezone';
 import { Button, FormControl, Select, Stack } from '@chakra-ui/core';
 
 import { ButtonTag, Search } from './';
-import { Tag } from '../helpers';
+import { State, Tag } from '../helpers';
 
 type Filter = {
-  filters: { [key: string]: Tag[] };
-  setSearch(search: string[]): void;
-  setTimezone(timezone: string): void;
-  toggleTag(filter: string, value: string, checked: boolean): void;
-  timezone: string;
+  setSearch: (search: string[]) => void;
+  setTimezone: (timezone: string) => void;
+  state: State;
+  toggleTag: (filter: string, value: string, checked: boolean) => void;
 };
 
-export function Filter({
-  filters,
-  setSearch,
-  setTimezone,
-  toggleTag,
-  timezone
-}: Filter) {
+export function Filter({ setSearch, setTimezone, state, toggleTag }: Filter) {
   const [open, setOpen] = useState(false);
   return (
     <Stack spacing={{ xs: 3, md: 6 }}>
       <FormControl>
-        <Search setSearch={setSearch} />
+        <Search search={state.search} setSearch={setSearch} />
       </FormControl>
       <Stack
         d={{ xs: open ? 'block' : 'none', md: 'block' }}
         spacing={{ xs: 3, md: 6 }}
       >
-        {Object.keys(filters).map((filter: string, index: number) => (
+        {Object.keys(state.filters).map((filter: string, index: number) => (
           <FormControl key={index}>
-            {filters[filter].map((tag: Tag, index: number) => (
+            {state.filters[filter].map((tag: Tag, index: number) => (
               <ButtonTag
                 filter={filter}
                 key={index}
@@ -50,7 +43,7 @@ export function Filter({
             onChange={(e: React.FormEvent<HTMLSelectElement>) =>
               setTimezone(e.currentTarget.value)
             }
-            value={timezone}
+            value={state.timezone}
           >
             {moment.tz.names().map((name: string, index: number) => (
               <option key={index}>{name}</option>
