+++ src/components/Meeting.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import { Box, Divider, Heading, Text, Tag } from '@chakra-ui/core';
+import { Box, Heading, Stack, Text, Tag } from '@chakra-ui/core';
 import Highlighter from 'react-highlight-words';
 import { Moment } from 'moment-timezone';
-// @ts-ignore
 import Linkify from 'react-linkify';
 
 import { ButtonPrimary } from './';
@@ -39,66 +38,86 @@ export function Meeting({
       rounded="md"
       shadow="md"
     >
-      <Box alignItems="baseline" mb={2}>
-        <Heading d={{ lg: 'inline' }} fontSize="2xl">
-          <Highlighter searchWords={search} textToHighlight={meeting.name} />
-        </Heading>
-        <Heading
-          as="h3"
-          color="gray.600"
-          d={{ lg: 'inline' }}
-          fontSize="lg"
-          fontWeight="normal"
-          ml={{ lg: 2 }}
-        >
-          {!meeting.time ? 'Ongoing' : meeting.time.format('dddd h:mm a')}
-        </Heading>
-      </Box>
-      {!!meeting.tags.length &&
-        meeting.tags.map((tag: string, index: number) => (
-          <Tag
-            bg={tags.includes(tag) ? 'gray.300' : 'gray.100'}
-            color={tags.includes(tag) ? 'gray.700' : 'gray.600'}
-            key={index}
-            mr={2}
-            my={2}
-            size="sm"
+      <Stack spacing={3}>
+        <Box alignItems="baseline">
+          <Heading d={{ lg: 'inline' }} fontSize="2xl">
+            <Highlighter searchWords={search} textToHighlight={meeting.name} />
+          </Heading>
+          <Heading
+            as="h3"
+            color="gray.600"
+            d={{ lg: 'inline' }}
+            fontSize="lg"
+            fontWeight="normal"
+            ml={{ lg: 2 }}
           >
-            {tag}
-          </Tag>
-        ))}
-      <Divider mb={0} role="separator" />
-      {!!meeting.url && (
-        <ButtonPrimary
-          icon="link"
-          link={meeting.url}
-          text={new URL(meeting.url).hostname.replace('www.', '')}
-          title={'Visit ' + meeting.url}
-        />
-      )}
-      {!!meeting.phone && (
-        <ButtonPrimary
-          icon="phone"
-          link={'tel:' + meeting.phone}
-          text="Phone"
-          title={'Call ' + meeting.phone}
-        />
-      )}
-      {!!meeting.email && (
-        <ButtonPrimary
-          icon="email"
-          link={'mailto:' + meeting.email}
-          text="Email"
-          title={'Email ' + meeting.email}
-        />
-      )}
-      <Divider mt={4} role="separator" />
-      {!!meeting.notes.length &&
-        meeting.notes.map((paragraph: string, key: number) => (
-          <Text key={key} mt={2}>
-            <Linkify>{paragraph}</Linkify>
-          </Text>
-        ))}
+            {!meeting.time ? 'Ongoing' : meeting.time.format('dddd h:mm a')}
+          </Heading>
+        </Box>
+        <Box>
+          {!!meeting.url && (
+            <Box float="left" mr={2} my={1}>
+              <ButtonPrimary
+                icon="link"
+                onClick={() => {
+                  window.open(meeting.url, '_blank');
+                }}
+                text={new URL(meeting.url).hostname.replace('www.', '')}
+                title={'Visit ' + meeting.url}
+              />
+            </Box>
+          )}
+          {!!meeting.phone && (
+            <Box float="left" mr={2} my={1}>
+              <ButtonPrimary
+                icon="phone"
+                onClick={() => {
+                  window.open('tel:' + meeting.phone, '_blank');
+                }}
+                text="Phone"
+                title={'Call ' + meeting.phone}
+              />
+            </Box>
+          )}
+          {!!meeting.email && (
+            <Box float="left" mr={2} my={1}>
+              <ButtonPrimary
+                icon="email"
+                onClick={() => {
+                  window.open('mailto:' + meeting.email, '_blank');
+                }}
+                text="Email"
+                title={'Email ' + meeting.email}
+              />
+            </Box>
+          )}
+        </Box>
+        {!!meeting.notes.length && (
+          <Stack spacing={2}>
+            {meeting.notes.map((paragraph: string, key: number) => (
+              <Text key={key}>
+                <Linkify>{paragraph}</Linkify>
+              </Text>
+            ))}
+          </Stack>
+        )}
+        {!!meeting.tags.length && (
+          <Box>
+            {meeting.tags.map((tag: string, index: number) => (
+              <Tag
+                bg={tags.includes(tag) ? 'gray.300' : 'gray.100'}
+                color={tags.includes(tag) ? 'gray.700' : 'gray.600'}
+                key={index}
+                mr={2}
+                my={1}
+                size="sm"
+              >
+                {tag}
+              </Tag>
+            ))}
+          </Box>
+        )}
+      </Stack>
     </Box>
   );
 }
