+++ src/components/index.ts
@@ -3,4 +3,5 @@ export * from './ButtonTag';
 export * from './Filter';
 export * from './Loading';
 export * from './Meeting';
+export * from './NoResults';
 export * from './Search';
