+++ src/helpers/index.ts
@@ -1,3 +1,3 @@
 export * from './data';
-export * from './google';
+export * from './filter';
 export * from './query';
